requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
alembic>=1.12.1
loguru>=0.7.2
PyYAML>=6.0.1
google-generativeai>=0.3.2
Pillow>=10.0.0
